Restrict SQLAlchemy to versions below 2.0 in pre-commit contexts
SQLAlchemy just released version 2.0, which is incompatible with
third-party stubs files for typechecking. We use SQLAlchemy 1.4.x,
so this should not affect us, but pre-commit builds its own
virtual environment and installs latest version by default, which
now breaks on any fresh pre-commit environment initialization.

This caps SQLAlchemy to the last version before 2.0 until we are
ready to upgrade.